+++ Unit02/src/Line.java
@@ -5,7 +5,7 @@
 
 public class Line
 {
-	public static double getSlope( int x1, int y1, int x2, int y2 )
+	public static double getSlope( float x1, float y1, float x2, float y2 )
 	{
 		return ((y2 - y1) / (x2 - x1));
 	}
