+++ Unit02/src/Trap.java
@@ -6,7 +6,7 @@
 
 public class Trap
 {
-	public static double area( int base1, int base2, int height )
+	public static double area( float base1, float base2, float height )
 	{
 		return ((base1 + base2)*.5*height );
 	}
